% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot method for linreg objects}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg object}

\item{...}{Additional arguments (not used)}
}
\value{
A list containing two ggplot objects
}
\description{
Creates diagnostic plots using ggplot2
}
