% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Extract residuals from linreg object}
\usage{
\method{resid}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of residuals
}
\description{
Extract residuals from linreg object
}
