% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary method for linreg objects}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisible summary object (prints automatically)
}
\description{
Summary method for linreg objects
}
